refactor(whatsrust): extract media wrappers
